--- a/ethcore/Cargo.toml
+++ b/ethcore/Cargo.toml
@@ -17,22 +17,24 @@ env_logger = "0.3"
 rustc-serialize = "0.3"
 heapsize = "0.3"
 rust-crypto = "0.2.34"
-time = "0.1"
-ethcore-util = { path = "../util" }
-evmjit = { path = "../evmjit", optional = true }
-ethash = { path = "../ethash" }
 num_cpus = "0.2"
-clippy = { version = "0.0.79", optional = true}
 crossbeam = "0.2.9"
 lazy_static = "0.2"
+bloomchain = "0.1"
+rayon = "0.3.1"
+semver = "0.2"
+bit-set = "0.4"
+time = "0.1"
+evmjit = { path = "../evmjit", optional = true }
+clippy = { version = "0.0.79", optional = true}
+ethash = { path = "../ethash" }
+ethcore-util = { path = "../util" }
 ethcore-devtools = { path = "../devtools" }
 ethjson = { path = "../json" }
-bloomchain = "0.1"
 ethcore-ipc = { path = "../ipc/rpc" }
-rayon = "0.3.1"
 ethstore = { path = "../ethstore" }
-semver = "0.2"
 ethcore-ipc-nano = { path = "../ipc/nano" }
+
 
 [dependencies.hyper]
 git = "https://github.com/ethcore/hyper"