--- a/ethcore/src/evm/interpreter/gasometer.rs
+++ b/ethcore/src/evm/interpreter/gasometer.rs
@@ -37,6 +37,7 @@ enum InstructionCost<Cost: CostType> {
 
 pub struct Gasometer<Gas: CostType> {
 	pub current_gas: Gas,
+	pub current_mem_gas: Gas,
 }
 
 impl<Gas: CostType> Gasometer<Gas> {
@@ -44,6 +45,7 @@ impl<Gas: CostType> Gasometer<Gas> {
 	pub fn new(current_gas: Gas) -> Self {
 		Gasometer {
 			current_gas: current_gas,
+			current_mem_gas: Gas::from(0),
 		}
 	}
 
@@ -62,7 +64,7 @@ impl<Gas: CostType> Gasometer<Gas> {
 		info: &InstructionInfo,
 		stack: &Stack<U256>,
 		current_mem_size: usize,
-	) -> evm::Result<(Gas, usize)> {
+	) -> evm::Result<(Gas, Gas, usize)> {
 		let schedule = ext.schedule();
 		let tier = instructions::get_tier_idx(info.tier);
 		let default_gas = Gas::from(schedule.tier_step_gas[tier]);
@@ -160,25 +162,25 @@ impl<Gas: CostType> Gasometer<Gas> {
 
 		match cost {
 			InstructionCost::Gas(gas) => {
-				Ok((gas, 0))
+				Ok((gas, self.current_mem_gas, 0))
 			},
 			InstructionCost::GasMem(gas, mem_size) => {
-				let (mem_gas, new_mem_size) = try!(self.mem_gas_cost(schedule, current_mem_size, &mem_size));
-				let gas = overflowing!(gas.overflow_add(mem_gas));
-				Ok((gas, new_mem_size))
+				let (mem_gas_cost, new_mem_gas, new_mem_size) = try!(self.mem_gas_cost(schedule, current_mem_size, &mem_size));
+				let gas = overflowing!(gas.overflow_add(mem_gas_cost));
+				Ok((gas, new_mem_gas, new_mem_size))
 			},
 			InstructionCost::GasMemCopy(gas, mem_size, copy) => {
-				let (mem_gas, new_mem_size) = try!(self.mem_gas_cost(schedule, current_mem_size, &mem_size));
+				let (mem_gas_cost, new_mem_gas, new_mem_size) = try!(self.mem_gas_cost(schedule, current_mem_size, &mem_size));
 				let copy = overflowing!(add_gas_usize(copy, 31)) >> 5;
 				let copy_gas = Gas::from(schedule.copy_gas) * copy;
 				let gas = overflowing!(gas.overflow_add(copy_gas));
-				let gas = overflowing!(gas.overflow_add(mem_gas));
-				Ok((gas, new_mem_size))
+				let gas = overflowing!(gas.overflow_add(mem_gas_cost));
+				Ok((gas, new_mem_gas, new_mem_size))
 			}
 		}
 	}
 
-	fn mem_gas_cost(&self, schedule: &evm::Schedule, current_mem_size: usize, mem_size: &Gas) -> evm::Result<(Gas, usize)> {
+	fn mem_gas_cost(&self, schedule: &evm::Schedule, current_mem_size: usize, mem_size: &Gas) -> evm::Result<(Gas, Gas, usize)> {
 		let gas_for_mem = |mem_size: Gas| {
 			let s = mem_size >> 5;
 			// s * memory_gas + s * s / quad_coeff_div
@@ -193,15 +195,14 @@ impl<Gas: CostType> Gasometer<Gas> {
 		let current_mem_size = Gas::from(current_mem_size);
 		let req_mem_size_rounded = (overflowing!(mem_size.overflow_add(Gas::from(31 as usize))) >> 5) << 5;
 
-		let mem_gas_cost = if req_mem_size_rounded > current_mem_size {
+		let (mem_gas_cost, new_mem_gas) = if req_mem_size_rounded > current_mem_size {
 			let new_mem_gas = try!(gas_for_mem(req_mem_size_rounded));
-			let current_mem_gas = try!(gas_for_mem(current_mem_size));
-			new_mem_gas - current_mem_gas
+			(new_mem_gas - self.current_mem_gas, new_mem_gas)
 		} else {
-			Gas::from(0)
+			(Gas::from(0), self.current_mem_gas)
 		};
 
-		Ok((mem_gas_cost, req_mem_size_rounded.as_usize()))
+		Ok((mem_gas_cost, new_mem_gas, req_mem_size_rounded.as_usize()))
 	}
 }
 
@@ -251,9 +252,10 @@ fn test_calculate_mem_cost() {
 	let mem_size = 5;
 
 	// when
-	let (mem_cost, mem_size) = gasometer.mem_gas_cost(&schedule, current_mem_size, &mem_size).unwrap();
+	let (mem_cost, new_mem_gas, mem_size) = gasometer.mem_gas_cost(&schedule, current_mem_size, &mem_size).unwrap();
 
 	// then
 	assert_eq!(mem_cost, 3);
+	assert_eq!(new_mem_gas, 3);
 	assert_eq!(mem_size, 32);
 }