--- a/ethcore/src/lib.rs
+++ b/ethcore/src/lib.rs
@@ -95,6 +95,7 @@ pub extern crate ethstore;
 extern crate semver;
 extern crate ethcore_ipc_nano as nanoipc;
 extern crate ethcore_devtools as devtools;
+extern crate bit_set;
 
 #[cfg(feature = "jit" )] extern crate evmjit;