1. Using shift instead of division for `quad_coeff_div`
2. Memoizing current_memory_gas
3. jump destinations optimization (using BitSet instead of HashSet)

Related #491 

Benchmarks (for all alterations): 
```
Loop
usize 0.344s vs 0.369s
U256 0.490s vs 0.517s
RNG
usize 1.292s vs 1.341s
U256 1.917s vs 1.923s
```

doesn't seem very conclusive - I will do some profiling during sync.